Move validator trait into new namespace
